+++ src/tooling/piral-cli/src/apps/new-pilet.ts
@@ -31,6 +31,8 @@ import {
   ensure,
   getCertificate,
   getAgent,
+  defaultImportmapVersions,
+  defaultRemoteTypesTarget,
 } from '../common';
 
 export interface NewPiletOptions {
@@ -178,6 +180,7 @@ export async function newPilet(baseDir = process.cwd(), options: NewPiletOptions
           peerDependencies: {},
           scripts: {},
           main: 'dist/index.js',
+          types: 'dist/index.d.ts',
           files: ['dist'],
         },
         undefined,
@@ -206,6 +209,8 @@ always-auth=true`,
         {
           $schema: piletJsonSchemaUrl,
           schemaVersion: defaultSchemaVersion,
+          importmapVersions: defaultImportmapVersions,
+          remoteTypesTarget: defaultRemoteTypesTarget,
           piralInstances: {},
         },
         undefined,