+++ src/tooling/piral-cli/src/npm-clients/rush.ts
@@ -8,10 +8,16 @@ import { jju } from '../external';
 // Helpers:
 const rushJson = 'rush.json';
 
-function runRushProcess(args: Array<string>, target: string, output?: NodeJS.WritableStream) {
+async function runRushProcess(args: Array<string>, target: string, output?: MemoryStream) {
   log('generalDebug_0003', 'Starting the Rush process ...');
   const cwd = resolve(process.cwd(), target);
-  return runCommand('rush', args, cwd, output);
+
+  try {
+    return await runCommand('rush', args, cwd, output);
+  } catch (err) {
+    log('generalInfo_0000', output.value || `rush failed due to ${err}`);
+    throw err;
+  }
 }
 
 function convert(flags: Array<string>) {